requirements.txt
Jinja2==3.1.4 
MarkupSafe==2.1.5 
Werkzeug==3.0.3 
blinker==1.8.2 
click==8.1.7 
flask==3.0.3 
itsdangerous==2.2.0
Flask-SQLAlchemy==3.1.1 
sqlalchemy==2.0.31 
typing-extensions==4.12.2
certifi==2024.7.4 
charset-normalizer==3.3.2 
idna==3.7 
requests==2.32.3 
urllib3==2.2.2
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==4.4.0 
distro==1.9.0 
h11==0.14.0 
httpcore==1.0.5 
httpx==0.27.0 
openai==0.28.0
pydantic==2.8.2 
pydantic-core==2.20.1 
sniffio==1.3.1 
tqdm==4.66.4
sqlitecloud==0.0.78